remove old notebook
