--- notebook-api/Dockerfile
@@ -1,27 +1,0 @@
-FROM alpine:3.8
-MAINTAINER Hail Team <hail@broadinstitute.org>
-
-RUN apk add \
-  bash \
-  gcc \
-  libffi-dev \
-  musl-dev \
-  openssl-dev \
-  python3 \
-  python3-dev && \
-  pip3 install -U pip==18.1 && \
-  pip3 install --no-cache-dir \
-  flask \
-  Flask_Sockets \
-  kubernetes \
-  'urllib3<1.24' \
-  ujson \
-  python-dotenv
-
-COPY notebook /notebook
-COPY notebook-worker-images /notebook
-
-EXPOSE 5000
-
-WORKDIR /notebook
-ENTRYPOINT ["python3", "notebook.py"]
