--- notebook-api/environment.yml
@@ -1,11 +1,0 @@
-name: notebook
-dependencies:
-  - python=3.7
-  - requests
-  - flask
-  - pip
-  - ujson
-  - pip:
-      - kubernetes
-      - Flask_Sockets
-      - python-dotenv
