--- notebook/.gitignore
@@ -1,4 +1,0 @@
-/deployment.yaml
-/notebook-image
-/notebook-worker-image
-
