--- notebook/SECRETS.md
@@ -1,9 +1,0 @@
-# Secrets
-
-name: `notebook-secrets`
-contents:
- - `secret-key` a random string
- - `password` a random string
- - `admin-password` a random string
-
-each should be one line, it may or may not be terminated by a newline.
