+++ notebook/environment.yml
@@ -1,9 +1,11 @@
 name: notebook
 dependencies:
-- python=3.7
-- requests
-- flask
-- pip
-- pip:
-  - kubernetes
-  - Flask_Sockets
+  - python=3.7
+  - requests
+  - flask
+  - pip
+  - ujson
+  - pip:
+      - kubernetes
+      - Flask_Sockets
+      - python-dotenv
