--- notebook/hail-ci-deploy.sh
@@ -1,9 +1,0 @@
-#!/bin/sh
-set -ex
-
-gcloud -q auth activate-service-account \
-  --key-file=/secrets/gcr-push-service-account-key.json
-
-gcloud -q auth configure-docker
-
-CI_BUILD=true make deploy push-all-latest-tag
