--- notebook/notebook/templates/admin-login.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>Hail Notebooks</title>
-    <link rel="stylesheet" type="text/css" href="/static/style.css">
-    <link rel="stylesheet" type="text/css" href="/static/centering.css">
-  </head>
-  <body>
-    <div class="center">
-      <h1>Hail Notebooks Admin Login</h1>
-      <form action="{{ form_action_url }}", method="post">
-        <p>
-          <label for="password">Admin Password:</label>
-          <input type="password" id="password" name="password" />
-        </p>
-        <input type="submit" value="Start a Hail Notebook" />
-      </form>
-    </div>
-  </body>
-</html>
