--- notebook/notebook/templates/index.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>Hail Notebooks</title>
-    <link rel="stylesheet" type="text/css" href="/static/style.css">
-    <link rel="stylesheet" type="text/css" href="/static/centering.css">
-  </head>
-  <body>
-    <div class="center">
-      <h1>Hail Notebooks</h1>
-      <form action="{{ form_action_url }}" method="post">
-        <p>
-          <label for="password">Class Key:</label>
-          <input type="text" name="password" id="password" />
-        </p>
-        <p>
-          <label for="image">Image:</label>
-          <select name="image" id="image">
-            {% for image in images %}
-            <option value="{{ image }}"{{'selected' if image == default else ''}}>{{ image }}</option>
-            {% endfor %}
-          </select>
-        </p>
-        <input type="submit" value="Start a Hail Notebook" />
-      </form>
-    </div>
-  </body>
-</html>
