--- notebook/notebook/templates/wait.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>Hail Notebooks</title>
-    <link rel="stylesheet" type="text/css" href="/static/style.css">
-    <link rel="stylesheet" type="text/css" href="/static/centering.css">
-    <script>
-      let protocol = location.protocol.replace("http", "ws")
-      let sock = new WebSocket(protocol + location.host + "/wait");
-      sock.onmessage = function (event) {
-        console.log(event)
-        window.location.replace(event.data)
-      }
-    </script>
-    <style>
-      .loader {
-        border: 16px solid #f3f3f3; /* Light grey */
-        border-top: 16px solid #3498db; /* Blue */
-        border-radius: 50%;
-        width: 120px;
-        height: 120px;
-        animation: spin 2s linear infinite;
-        margin-left: auto;
-        margin-right: auto;
-      }
-
-      @keyframes spin {
-        0% { transform: rotate(0deg); }
-        100% { transform: rotate(360deg); }
-      }
-    </style>
-  </head>
-  <body>
-    <div class="center">
-      <h1>Waiting for Hail Notebook to Start</h1>
-      <div class="loader" />
-    </div>
-  </body>
-</html>
