--- notebook/notebook/templates/workers.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>Hail Notebooks</title>
-    <link rel="stylesheet" type="text/css" href="/static/style.css">
-  </head>
-  <body>
-    <div class="center">
-      <h1>Hail Notebook Workers</h1>
-      <p>Leader instance: {{ leader_instance }}</p>
-      <form action="{{ workers_url }}/delete-all-workers", method="post">
-        <input type="submit" value="Kill all workers" />
-      </form>
-      <table>
-        <thead>
-          <tr>
-            <th align="left">name</th>
-            <th align="left">status</th>
-            <th align="left">status explanation</th>
-            <th align="left">start time</th>
-            <th align="left">service name</th>
-            <th align="left">created by instance id</th>
-            <th align="left">delete</th>
-          </tr>
-        </thead>
-        <tbody>
-          {% for w in workers %}
-          <tr>
-            <td>{{ w[0].metadata.name }}</td>
-            <td>{{ w[0].status.phase }}</td>
-            <td>{{ w[0].status.message }}</td>
-            <td>{{ w[0].status.start_time }}</td>
-            <td>{{ w[1].metadata.name }}</td>
-            <td>{{ w[0].metadata.labels.get("hail.is/notebook-instance") }}</td>
-            <td><a href="{{ workers_url }}/{{ w[0].metadata.name }}/{{w[1].metadata.name}}/delete">delete</a></td>
-          </tr>
-          {% endfor %}
-        </tbody>
-      </table>
-    </div>
-  </body>
-</html>
